requirements.txt
imutils==0.5.3
numpy==1.18.5
scipy==1.4.1
tensorflow-gpu==2.1.2
opencv-python==4.5.4.58
pyyaml==5.4.1 